requirements.txt
aiofiles==23.2.1
aiogram==3.11.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
attrs==24.2.0
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.7
frozenlist==1.5.0
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
magic-filter==1.0.12
MarkupSafe==3.0.2
multidict==6.1.0
propcache==0.2.1
psycopg2-binary==2.9.10
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
python-multipart==0.0.19
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
starlette-admin==0.14.1
typing_extensions==4.12.2
uvicorn==0.30.6
yarl==1.18.3
